pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.22",
    "httpx>=0.28.1",
    "sse-starlette>=3.1.2",
    "starlette>=0.51.0",
    "uvicorn>=0.40.0",
]
